requirements.txt
# Production dependencies for Shopee Webhook Receiver
fastapi>=0.104.0
uvicorn[standard]>=0.24.0
python-dotenv>=1.0.0
requests>=2.31.0
pydantic>=2.5.0
pydantic-settings>=2.0.0

# HTTP Client
httpx>=0.25.0

# Error Monitoring
sentry-sdk[fastapi]>=1.40.0

# Google Sheets (for processor)
gspread>=5.12.0
google-auth>=2.23.0
google-auth-oauthlib>=1.1.0
google-auth-httplib2>=0.1.1
google-api-python-client>=2.100.0
